Initial dependency install for webpack to utilize CSS Modules.
