+++ __mocks__/axios.js
@@ -1,0 +1,2 @@
+import mockAxios from 'jest-mock-axios';
+export default mockAxios;
