+++ client/assets/style.css
@@ -1,0 +1,333 @@
+/* * {
+  font-family: Arial, Helvetica, sans-serif;
+} */
+body {
+  font-family: Arial, Helvetica, sans-serif;
+  background-color: white;
+  max-width: 100%;
+  overflow-x: hidden;
+  margin: 0px;
+  box-sizing: border-box;
+  align-content: center;
+  line-height: 1.4285;
+}
+
+div {
+  display: block;
+}
+
+li {
+  display: list-item;
+  /* margin-left: 20px; */
+}
+
+hr {
+  margin-top: 16px;
+  margin-right: 32px;
+  /* margin-bottom: 16px; */
+  border-width: 1px 0 0;
+  border-top-style: solid;
+  border-top-color: rgb(214, 214, 214);
+}
+
+/* CONTAINER */
+
+#mainContainer {
+  padding: 16px;
+  max-width: 1400px;
+  background-color: rgb(247, 247, 247);
+  /* margin-top: 16px;
+  margin-bottom: 16px; */
+  margin: 16px auto;
+}
+
+.marginLeft {
+  margin-left: 16px;
+  margin-right: 16px;
+  font-size: 14px;
+  flex: 0 1 auto;
+  flex-flow: row wrap;
+  height: 100%;
+  width: 100%;
+}
+
+.container-white {
+  margin: 16px;
+  padding: 4px;
+  color: rgb(51, 51 ,51);
+  background-color: white;
+}
+
+.container-highlights {
+  box-sizing: border-box;
+  margin: 16px;
+  padding: 0;
+  display: flex;
+  flex-flow: row wrap;
+  height: 100%;
+  width: 100%;
+  list-style: outside none none;
+}
+
+/* HEADING */
+
+#moduleTitle {
+  font-size: 23px;
+  margin-bottom: 8px;
+  text-align: center;
+}
+
+#tabHeading {
+  width: 100%;
+  border-bottom: 1px solid rgb(136, 136, 136);
+  text-size-adjust: 100%;
+}
+
+#tab-list {
+  border-spacing: 40px 0px;
+  vertical-align: baseline;
+  padding: 4px 1px 0;
+  display: table;
+  list-style: outside none none;
+  margin: 0px auto;
+
+}
+
+/* TABS */
+
+.tab {
+  display: table-cell;
+  color: #333333;
+  padding: 8px 12px;
+  text-align: center;
+  font-size: 16px;
+  color: rgb(102, 102, 102);
+  cursor: pointer;
+}
+
+.tab.selected {
+  transition: border-color 200ms ease 0ms,
+              color 200ms ease 0ms;
+  border-bottom: 3px solid rgb(204, 0, 0);
+  color: rgb(51, 51, 51);
+  font-weight: bold;
+
+}
+
+/* BUTTONS */
+
+.btn {
+  /* appearance: button; */
+  content: "";
+  cursor: pointer;
+  display: inline-block;
+  /* position: absolute; */
+  /* border: 5px solid transparent; */
+  border-radius: 4px;
+  border-style: solid;
+  border-width: 1px;
+  border-color: rgb(136, 136, 136);
+  background-color: rgb(255, 255, 255);
+  color: rgb(51, 51, 51);
+  margin: 0px;
+  /* opacity: 0; */
+  padding: 6px 12px;
+  /* top: -2px;
+  right: -2px;
+  bottom: -2px;
+  left: -2px; */
+  /* animation-duration: 500ms; */
+  text-align: center;
+  font-size: 12px;
+  /* align-items: flex-start; */
+}
+
+.btn:hover {
+  color: rgb(51, 51, 51);
+  border-color: rgb(51, 51, 51);
+  background-color: rgb(244, 244, 244);
+}
+
+.btn:active {
+  color: rgb(255, 255, 255);
+  background-color: rgb(51, 51, 51);
+  border-color: rgb(51, 51, 51);
+}
+
+.btn.question {
+  width: 100%;
+  padding: 12px 16px;
+  font-size: 16px;
+}
+
+.btn.question.ask {
+  color: white;
+  background-color: rgb(204, 0, 0);
+  border-color: rgb(204, 0, 0);
+}
+
+.btn.question.ask:hover {
+  color: rgb(255, 255, 255);
+  background-color: rgb(170, 0, 0);
+  border-color: rgb(170, 0, 0);
+}
+
+.btn.question.ask:active {
+  color: rgb(255, 255, 255);
+  background-color: rgb(121, 0, 0);
+  border-color: rgb(121, 0, 0);
+}
+
+/* SHIPPING & RETURNS */
+
+/* #shipping-returns {
+  margin: 16px;
+  padding: 4px;
+} */
+
+.shipping-green {
+  color: rgb(0, 102, 1);
+}
+
+.shipping-warning {
+  color: rgb(232, 105, 0);
+}
+
+/* HIGHLIGHTS */
+
+.highlight-element {
+  /* flex: 0 0 auto; */
+  flex-grow: 1;
+  display: flex;
+  padding-right: 16px;
+  padding-top: 12px;
+  /* margin-left: initial; */
+  flex-basis: 23%;
+  max-width: 23%;
+}
+
+.highlight-item {
+  list-style-type: disc;
+  padding-right: 16px;
+  margin-right: 16px;
+}
+
+/* SPECIFICATIONS AND DESCRIPTION */
+
+.container-spec-desc {
+  display: flex;
+  flex: 0 1 auto;
+  flex-flow: row wrap;
+}
+
+.collapsed {
+  overflow: hidden;
+  height: 10px;
+  opacity: .9;
+  background-image: linear-gradient(to bottom, rgb(247, 247, 247), white)
+}
+
+.column {
+  /* padding-left: 16px;
+  padding-right: 16px; */
+  flex-basis: 50%;
+  max-width: 50%;
+  flex-direction: row;
+  font-size: 14px;
+}
+
+.report-details {
+  padding-top: 16px;
+  padding-bottom: 32px;
+}
+
+.description-container {
+  padding-right: 16px;
+  padding-bottom: 16px;
+}
+
+.spec-container {
+  padding-right: 32px;
+  padding-bottom: 16px;
+}
+
+/* Q & A */
+
+.container-questions {
+  padding-top: 16px;
+}
+
+.question-row {
+  font-size: 14px;
+  font-weight: bold;
+}
+
+.question-row-name {
+  font-size: 12px;
+  color: rgb(102, 102, 102);
+  margin-top: -8px;
+  margin-bottom: 16px;
+}
+
+.question-button-row {
+  /* align-self: center; */
+  /* position: relative; */
+  display: flex;
+  flex-direction: row;
+  flex-grow: 1;
+  /* width: 95%; */
+  /* max-width: 50%; */
+  flex-basis: 50%;
+  /* flex: 1 0 auto; */
+  padding-left: 8px;
+  padding-right: 8px;
+  justify-content: space-evenly;
+}
+
+.question-button-column {
+  flex-basis: 50%;
+  max-width: 50%;
+  font-size: 16px;
+  padding-left: 8px;
+  padding-right: 8px;
+}
+
+.container-answer {
+  padding-left: 16px;
+  /* padding-bottom: 8px; */
+  padding-right: 16px;
+}
+
+.answer-bold {
+  font-size: 16px;
+  font-weight: bold;
+}
+
+.answer-row-name {
+  font-size: 12px;
+  color: rgb(102, 102, 102);
+  margin-bottom: 16px;
+  margin-top: 4px;
+}
+
+.answer-click {
+  margin-left: 16px;
+}
+
+/* .helpful {
+  margin-left: 16px;
+}
+
+.not-helpful {
+  margin-left: 16px;
+}
+
+.report-answer {
+  margin-left: 16px;
+} */
+
+/* DISPLAY */
+
+.display-none {
+  display: none !important;
+}
