+++ client/setupTests.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
+configure({ adapter: new Adapter() });
