+++ package.json
@@ -7,20 +7,22 @@
   "private": true,
   "jest": {
     "reporters": [
-      [
-        "jest-nyancat-reporter",
-        {
-          "suppressErrorReporter": false
-        }
-      ]
+      "default"
     ],
-    "testEnvironment": "node"
+    "testEnvironment": "node",
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ],
+    "setupFiles": [
+      "./client/setupTests.js"
+    ]
   },
   "engines": {
     "node": ">=6.13.0"
   },
   "scripts": {
     "test": "jest",
+    "test-coverage": "jest --coverage",
     "test:watch": "jest --watch",
     "start": "node server/app.js",
     "build": "webpack --mode=production",
@@ -36,8 +38,8 @@
     "mariadb": "^2.5.1",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.32",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^16.0.1",
+    "react-dom": "^16.0.1",
     "react-moment": "^1.0.0",
     "sequelize": "^6.3.5"
   },
@@ -45,13 +47,21 @@
     "@babel/core": "^7.12.3",
     "@babel/preset-env": "^7.12.1",
     "@babel/preset-react": "^7.12.5",
-    "babel-polyfill": "^6.26.0",
+    "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.1",
+    "babel-polyfill": "^6.26.0",
+    "css-loader": "^5.0.1",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
+    "enzyme-to-json": "^3.6.1",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "faker": "^5.1.0",
     "jest": "^26.6.3",
+    "jest-mock-axios": "^4.2.1",
     "jest-nyancat-reporter": "^2.0.0",
     "nodemon": "^2.0.6",
+    "react-test-renderer": "^16.0.1",
+    "style-loader": "^2.0.0",
     "supertest": "^6.0.1",
     "webpack": "^5.4.0",
     "webpack-cli": "^4.2.0"
