+++ public/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="./assets/style.css" rel="stylesheet">
+    <!-- <link href="./assets/style.css" rel="stylesheet"> -->
     <title>About This Item</title>
   </head>
   <body>
